lambda handlers separated in codebase
